# openapi.yaml
openapi: "3.0.3"
info:
  title: gr33n API
  version: "0.2.0"
  description: |
    Local-first automated agriculture platform API.
    All endpoints run on localhost:8080 by default.

    **Authentication:**
    - Public routes require no auth.
    - Pi routes require `X-API-Key` header.
    - Dashboard routes require `Authorization: Bearer <JWT>`.
    - In `AUTH_MODE=dev` (dev-tagged build only), auth is bypassed.
    - `AUTH_MODE=auth_test` enforces the same auth as production on a dev-tagged build only.

    **Operator docs:** Phase 13 index — `docs/phase-13-operator-documentation.md`. **Phase 14 index** — `docs/phase-14-operator-documentation.md`. Farm audit trail — `docs/audit-events-operator-playbook.md`. Insert Commons pipeline — `docs/insert-commons-pipeline-runbook.md`; receiver — `docs/insert-commons-receiver-playbook.md`. Commons catalog (gr33n_inserts) — `docs/commons-catalog-operator-playbook.md`. Receipt storage cutover — `docs/receipt-storage-cutover-runbook.md`. Push alerts (FCM) — `docs/notifications-operator-playbook.md`. Domain module stubs (DB only; no API yet) — `docs/domain-modules-operator-playbook.md`. Optional mobile wrapper — `docs/mobile-distribution.md`.

    **Phase 13 API areas:** `organizations` (multi-farm grouping, usage summary), `farms` (Insert Commons opt-in/sync/history, farm-scoped audit events), `costs` / attachments (finance mappings, exports, receipts).
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0
servers:
...
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: Login
      description: Authenticate with username/password. Returns a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /auth/password:
    patch:
      tags: [auth]
      summary: Change password
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string, format: password }
                new_password: { type: string, format: password }
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/register:
    post:
      tags: [auth]
      summary: Register a new user
      description: Creates auth.users + profiles row. Invited users (null password) can set password here.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password, minLength: 8 }
                full_name: { type: string }
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  expires_in: { type: integer }
                  user_id: { type: string, format: uuid }
        "409":
          description: User already exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # ── Pi routes (API key) ────────────────────────────────────
  /sensors/{id}/readings:
    get:
      tags: [sensors]
      summary: List sensor readings in a time range
      description: |
        Returns readings from `since` through `until` (inclusive), oldest first.
        Defaults to the last 24 hours. At most `limit` rows (newest within the window).
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/ID" }
        - name: since
          in: query
          required: false
          schema: { type: string, format: date-time }
          description: Range start (RFC3339). Defaults to 24 hours before `until`.
        - name: until
          in: query
          required: false
          schema: { type: string, format: date-time }
          description: Range end (RFC3339). Defaults to now (UTC).
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 5000, default: 500 }
      responses:
        "200":
          description: Readings oldest-first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SensorReading" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [sensors]
      summary: Post a sensor reading (Pi client)
      description: Called by the Raspberry Pi client daemon. Triggers SSE push to connected dashboards.
      security: [{ apiKeyAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/ID" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SensorReadingCreate" }
            example:
              value_raw: 22.5
              is_valid: true
              battery_level_percent: 87.0
      responses:
        "201":
          description: Reading stored and normalized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SensorReading" }
  /sensors/readings/batch:
    post:
      tags: [sensors]
      summary: Post multiple sensor readings (edge / MQTT bridge)
      description: |
        One HTTP request for several readings (e.g. MQTT subscriber on a gateway).
        Inserts run in a single transaction. At most 64 items per request.
        Same `X-API-Key` as `POST /sensors/{id}/readings`.
      security: [{ apiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              maxItems: 64
              items: { $ref: "#/components/schemas/SensorReadingBatchItem" }
            example:
              - { sensor_id: 1, value_raw: 22.5, is_valid: true }
              - { sensor_id: 2, value_raw: 61.0 }
      responses:
        "201":
          description: All readings stored
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SensorReadingBatchResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /devices/{id}/status:
    patch:
      tags: [devices]
      summary: Update device status (Pi heartbeat)
      security: [{ apiKeyAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/ID" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: "#/components/schemas/DeviceStatusEnum" }
            example: { status: online }
      responses:
        "200":
          description: Updated device
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }

  # ── Units ──────────────────────────────────────────────────
  /units:
    get:
      tags: [units]
      summary: List all units of measure
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of units
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Unit" }

  # ── Commons catalog (gr33n_inserts) ───────────────────────
  /commons/catalog:
    get:
      tags: [commons]
      summary: List published commons catalog entries
      description: >
        Returns metadata for operator-curated packs (documentation, reference JSON, future insert templates).
        Response rows omit the heavy `body`; use `GET /commons/catalog/{slug}` for the full payload.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: q
          in: query
          schema: { type: string }
          description: Search title, summary, or tags (case-insensitive substring)
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 100 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: List of catalog entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CommonsCatalogListEntry" }
  /commons/catalog/{slug}:
    get:
      tags: [commons]
      summary: Get one published catalog entry by slug
      security: [{ bearerAuth: [] }]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Full entry including `body` (JSON)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommonsCatalogEntryDetail" }
        "404":
          $ref: "#/components/responses/NotFound"
  /farms/{id}/commons/catalog-imports:
    get:
      tags: [commons]
      summary: List catalog imports recorded for this farm
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/ID" }
      responses:
        "200":
          description: Import audit rows with catalog metadata
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FarmCommonsCatalogImportListRow" }
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [commons]
      summary: Record a catalog import for this farm (farm admin)
      description: >
        Idempotent per farm and catalog entry. Does not execute SQL or mutate farm data automatically —
        records provenance and returns the catalog `body` for clients, automation, or future workers.
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/ID" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CommonsCatalogImportRequest" }
      responses:
        "200":
          description: Import row plus catalog entry snapshot
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommonsCatalogImportResponse" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Farms ──────────────────────────────────────────────────
  /farms:
    get:
      tags: [farms]
      summary: List farms for a user
      security: [{ bearerAuth: [] }]
      parameters:
        - name: user_id
          in: query
          required: false
          schema: { type: string, format: uuid }
          description: Filter by user membership. Omit to list all farms.
      responses:
        "200":
          description: List of farms
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Farm" }
    post:
      tags: [farms]
      summary: Create a farm
      description: >
        If `bootstrap_template` is omitted and the farm is not linked to an organization with a default template,
        the response body is a plain Farm object.
        If `bootstrap_template` is present (including `none`), or if the farm has `organization_id` set and that
        organization has `default_bootstrap_template`, the response is `{ farm, bootstrap }` where `bootstrap`
        is the outcome of `gr33ncore.apply_farm_bootstrap_template` (skipped, applied, already_applied, or error detail).
        Sending `bootstrap_template: "none"` skips both an explicit pack and the organization default.
        Auto-bootstrap runs only when the authenticated user matches `owner_user_id`.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FarmCreate" }
      responses:
        "201":
          description: Created farm (or farm + bootstrap when bootstrap_template was sent)
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: "#/components/schemas/Farm" }
                  - { $ref: "#/components/schemas/FarmCreatedWithBootstrap" }
  /farms/{id}/bootstrap-template:
    post:
      tags: [farms]
      summary: Apply a versioned starter template to an existing farm (farm admin)
      description: >
        Idempotent per farm and template key. Currently supported `template` value: `jadam_indoor_photoperiod_v1`.
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/ID" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [template]
              properties:
                template:
                  type: string
                  example: jadam_indoor_photoperiod_v1
      responses:
        "200":
          description: Bootstrap result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FarmBootstrapApplyResponse" }
        "400":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /farms/{id}:
    get:
      tags: [farms]
      summary: Get farm by ID
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/ID" }
      responses:
        "200":
          description: Farm object
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Farm" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [farms]
      summary: Update a farm
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/ID" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FarmUpdate" }
      responses:
        "200":
          description: Updated farm
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Farm" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [farms]
      summary: Soft-delete a farm
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/ID" }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                updated_by_user_id: { type: string, format: uuid }
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
  /farms/{id}/organization:
    patch:
      tags: [farms]
      summary: Link or unlink farm from an organization (farm admin; org admin required to link)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/ID" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FarmOrganizationPatch" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Farm" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /organizations:
    post:
      tags: [organizations]
      summary: Create an organization (caller becomes owner)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrganizationCreate" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
    get:
      tags: [organizations]
      summary: List organizations the current user belongs to
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OrganizationWithMembership" }
  /organizations/{id}:
    get:
      tags: [organizations]
      summary: Get organization by id (members only)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/ID" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [organizations]
      summary: Update organization name and billing hooks (owner or org admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/ID" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrganizationUpdate" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /organizations/{id}/usage-summary:
    get:
      tags: [organizations]
      summary: Aggregated usage counts across farms in the organization (metering hook)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/ID" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrganizationUsageSummary" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /sensors/{id}:
    get:
      tags: [sensors]
      summary: Get sensor by ID
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/ID" }
      responses:
        "200":
          description: Sensor object
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sensor" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [sensors]
      summary: Soft-delete a sensor
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/ID" }
      responses:
        "204": { description: Deleted }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...